Bundle Data Workspace and SQL Projects with esbuild
