+++ extensions/data-workspace/package.json
@@ -18,7 +18,7 @@
         "onCommand:projects.openExisting",
         "onCommand:projects.new"
     ],
-    "main": "./out/src/main.js",
+    "main": "./dist/main.js",
     "l10n": "./l10n",
     "repository": {
         "type": "git",
@@ -33,17 +33,19 @@
     "scripts": {
         "build:runtime-localization": "node ../../scripts/localization-generate.js data-workspace .",
         "build:extension": "tsc -p tsconfig.json",
+        "build:extension-bundle": "node esbuild.mjs",
         "build:prepare": "npm run build:runtime-localization",
-        "build": "npm run build:prepare && npm run build:extension",
+        "build": "npm run build:prepare && npm run build:extension && npm run build:extension-bundle",
         "compile": "npm run build",
         "watch:extension": "tsc -watch -p tsconfig.json --preserveWatchOutput",
-        "watch": "npm run build:prepare && run-p -l watch:extension",
+        "watch:extension-bundle": "node esbuild.mjs --watch",
+        "watch": "npm run build:prepare && run-p -l watch:extension watch:extension-bundle",
         "lint": "eslint ./src ./test",
         "localization:extract": "node ../../scripts/localization-extract.js data-workspace",
         "localization:generate": "node ../../scripts/localization-generate.js data-workspace .",
         "localization": "npm run localization:extract",
         "test": "npm run build && vscode-test --coverage",
-        "package": "npm run build && vsce package"
+        "package": "npm run build && vsce package --no-dependencies"
     },
     "capabilities": {
         "virtualWorkspaces": false,
@@ -201,13 +203,12 @@
     },
     "dependencies": {
         "@microsoft/ads-extension-telemetry": "^3.0.1",
-        "fast-glob": "^3.2.7",
+        "glob": "^13.0.6",
         "vscode-nls": "^4.0.0"
     },
     "devDependencies": {
         "@istanbuljs/nyc-config-typescript": "^1.0.2",
         "@types/chai": "^5.2.2",
-        "@types/glob": "^8.1.0",
         "@types/mocha": "^5.2.7",
         "@types/node": "^22.13.14",
         "@types/sinon": "^10.0.12",
@@ -218,7 +219,7 @@
         "@vscode/test-electron": "^3.0.0",
         "@vscode/vsce": "^3.9.1",
         "chai": "^5.3.3",
-        "glob": "^13.0.6",
+        "esbuild": "^0.28.1",
         "mocha-junit-reporter": "^2.2.1",
         "mocha-multi-reporters": "^1.5.1",
         "npm-run-all": "^4.1.5",
