+++ extensions/mssql/scripts/bundle-extension.js
@@ -3,52 +3,30 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-const logger = require("../../../scripts/terminal-logger");
-const { esbuildProblemMatcherPlugin, build, watch } = require("./esbuild-utils");
-
-// Parse arguments
-const args = process.argv.slice(2);
-const isProd = args.includes("--prod") || args.includes("-p");
-const isWatch = args.includes("--watch") || args.includes("-w");
+const { createNodeExtensionConfig, run } = require("../../../scripts/esbuild-utils");
 
 // Build configuration
-const config = {
-    entryPoints: {
-        extension: "src/extension.ts",
-        serviceInstallerUtil: "src/languageservice/serviceInstallerUtil.ts",
-    },
-    bundle: true,
-    outdir: "dist",
-    platform: "node",
-    loader: {
-        ".ts": "ts",
-        ".js": "js",
-        ".json": "json",
-        ".node": "file",
-    },
-    tsconfig: "./tsconfig.extension.json",
-    plugins: [esbuildProblemMatcherPlugin("extension")],
-    nodePaths: ["./node_modules"],
-    sourcemap: !isProd,
-    sourcesContent: false,
-    metafile: !isProd,
-    external: ["vscode", "vscode-mssql"],
-    minify: isProd,
-};
-
-// Main execution
-async function main() {
-    if (isWatch) {
-        logger.header("Building extension (watch mode)");
-        await watch(config);
-    } else {
-        logger.header(`Building extension`);
-        const success = await build(config, isProd);
-        process.exit(success ? 0 : 1);
-    }
-}
-
-// Run if called directly
-if (require.main === module) {
-    main();
-}
+void run(
+    ({ isProd }) =>
+        createNodeExtensionConfig({
+            entryPoints: {
+                extension: "src/extension.ts",
+                serviceInstallerUtil: "src/languageservice/serviceInstallerUtil.ts",
+            },
+            external: ["vscode-mssql"],
+            loader: {
+                ".ts": "ts",
+                ".js": "js",
+                ".json": "json",
+                ".node": "file",
+            },
+            metafile: !isProd,
+            minify: isProd,
+            nodePaths: ["./node_modules"],
+            outdir: "dist",
+            sourcemap: !isProd,
+            sourcesContent: false,
+            tsconfig: "./tsconfig.extension.json",
+        }),
+    "extension",
+);
