+++ extensions/sql-database-projects/scripts/build.js
@@ -6,6 +6,7 @@ const execOptions = { stdio: "inherit", shell: process.platform === "win32" };
 try {
     execFileSync(npmCommand, ["run", "build:prepare"], execOptions);
     execFileSync(npmCommand, ["run", "build:extension"], execOptions);
+    execFileSync(npmCommand, ["run", "build:extension-bundle"], execOptions);
 } catch (error) {
     process.exit(1);
 }
