+++ extensions/sql-database-projects/src/controllers/mainController.ts
@@ -20,7 +20,7 @@ import { IconPathHelper } from "../common/iconHelper";
 import { WorkspaceTreeItem } from "dataworkspace";
 import * as constants from "../common/constants";
 import { SqlDatabaseProjectProvider } from "../projectProvider/projectProvider";
-import { ItemType } from "sqldbproj";
+import { ItemType } from "../sqldbproj";
 import { FileNode } from "../models/tree/fileFolderTreeItem";
 import { HttpClient } from "../http/httpClient";
