+++ extensions/sql-database-projects/src/projectProvider/projectProvider.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as dataworkspace from "dataworkspace";
-import * as sqldbproj from "sqldbproj";
+import * as sqldbproj from "../sqldbproj";
 import * as vscode from "vscode";
 import * as constants from "../common/constants";
 import { IconPathHelper } from "../common/iconHelper";
