+++ extensions/sql-database-projects/src/tools/buildHelper.ts
@@ -7,8 +7,8 @@ import * as vscode from "vscode";
 import * as path from "path";
 import { promises as fs } from "fs";
 import * as utils from "../common/utils";
-import * as sqldbproj from "sqldbproj";
-import * as extractZip from "extract-zip";
+import * as sqldbproj from "../sqldbproj";
+import extractZip from "extract-zip";
 import * as constants from "../common/constants";
 import { HttpClient } from "../http/httpClient";
 import { getMicrosoftBuildSqlVersion } from "./netcoreTool";
