+++ extensions/sql-database-projects/tsconfig.extension.json
@@ -1,5 +1,8 @@
 {
     "extends": "../tsconfig.base.json",
+    "compilerOptions": {
+        "esModuleInterop": true
+    },
     "include": ["src", "test", "typings"],
     "exclude": ["node_modules"]
 }
